Refactor asset tree structure by consolidating TreeNodeModel and related classes, and remove unused AssetTreeTile widget
