+++ pubspec.yaml
@@ -32,5 +32,4 @@ flutter:
   uses-material-design: true
   assets:
     - assets/
-    - assets/images/
     - assets/icons/
